+++ spec/ruby_coincheck_client_spec.rb
@@ -1,7 +1,9 @@
-require 'spec_helper'
+# frozen_string_literal: true
+
+require "spec_helper"
 
 describe RubyCoincheckClient do
-  it 'has a version number' do
-    expect(RubyCoincheckClient::VERSION).not_to be nil
+  it "has a version number" do
+    expect(RubyCoincheckClient::VERSION).to match(/\A\d+\.\d+\.\d+\z/)
   end
 end